Rename project name and isolate gradle modules(core and idea-plugin)
