+++ .github/workflows/build.yml
@@ -13,9 +13,9 @@ on:
       - '*.txt'
 
 jobs:
-  build:
+  setup:
     runs-on: ubuntu-latest
-    timeout-minutes: 10
+    timeout-minutes: 5
     steps:
       - name: Checkout repository
         uses: actions/checkout@v2
@@ -25,7 +25,35 @@ jobs:
           java-version: '11'
           distribution: 'adopt'
           cache: gradle
+
+  build-core:
+    needs: [ setup ]
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+    steps:
+      # Check out current repository
+      - name: Checkout repository
+        uses: actions/checkout@v2
+
+      # Validate gradle wrapper
       - name: Validate Gradle Wrapper
+        uses: gradle/wrapper-validation-action@v1
+
+      # Run tests
+      - name: Run tests for kotlinify-core
+        run: ./gradlew :kotlinify-core:test
+
+  build-idea-plugin:
+    needs: [ setup ]
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+    steps:
+      # Check out current repository
+      - name: Checkout repository
+        uses: actions/checkout@v2
+
+      # Validate gradle wrapper
+      - name: Validate gradle wrapper
         uses: gradle/wrapper-validation-action@v1
 
       # Set environment variables
@@ -40,8 +68,8 @@ jobs:
           ./gradlew listProductsReleases
 
       # Run tests
-      - name: Run tests
-        run: ./gradlew test
+      - name: Run tests for kotlinify-idea-plugin
+        run: ./gradlew :kotlinify-idea-plugin:test
 
       # Upload tests results
       - name: Upload test results
@@ -60,14 +88,14 @@ jobs:
 
       # Run the IntelliJ Plugin Verifier tool
       - name: Check the binary compatibility
-        run: ./gradlew runPluginVerifier -Pplugin.verifier.home.dir=${{ steps.properties.outputs.pluginVerifierHomeDir }}
+        run: ./gradlew :kotlinify-idea-plugin:runPluginVerifier -Pplugin.verifier.home.dir=${{ steps.properties.outputs.pluginVerifierHomeDir }}
 
       # Upload results of the IntelliJ Plugin Verifier
       - name: Upload IntelliJ Plugin Verifier results
         uses: actions/upload-artifact@v2
         with:
           name: pluginVerifier
-          path: ${{ github.workspace }}/build/reports/pluginVerifier
+          path: ${{ github.workspace }}/kotlinify-idea-plugin/build/reports/pluginVerifier
         if: ${{ always() }}
 
       # Prepare plugin archive content for creating artifact
@@ -75,7 +103,7 @@ jobs:
         id: artifact
         shell: bash
         run: |
-          cd ${{ github.workspace }}/build/distributions
+          cd ${{ github.workspace }}/kotlinify-idea-plugin/build/distributions
           FILENAME=`ls *.zip`
           unzip "$FILENAME" -d content
           echo "::set-output name=filename::${FILENAME:0:-4}"
@@ -85,4 +113,4 @@ jobs:
         uses: actions/upload-artifact@v2.2.4
         with:
           name: ${{ steps.artifact.outputs.filename }}
-          path: ./build/distributions/content/*/*
+          path: ${{ github.workspace }}/kotlinify-idea-plugin/build/distributions/content/*/*
