+++ .run/Build plugin.run.xml
@@ -1,19 +1,3 @@
-<!--
-  ~ Copyright (c) 2022 Tatsuya Maki
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~         http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Bundle the plugin as a distribution" type="GradleRunConfiguration" factoryName="Gradle">
     <ExternalSystemSettings>
@@ -26,7 +10,7 @@
       </option>
       <option name="taskNames">
         <list>
-          <option value="buildPlugin" />
+          <option value=":kotlinify-idea-plugin:buildPlugin" />
         </list>
       </option>
       <option name="vmOptions" />
