+++ .run/Run IDE.run.xml
@@ -1,19 +1,3 @@
-<!--
-  ~ Copyright (c) 2022 Tatsuya Maki
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~         http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run IDE with plugin" type="GradleRunConfiguration" factoryName="Gradle">
     <log_file alias="idea.log" path="$PROJECT_DIR$/build/idea-sandbox/system/log/idea.log" />
@@ -27,7 +11,7 @@
       </option>
       <option name="taskNames">
         <list>
-          <option value="runIde" />
+          <option value=":kotlinify-idea-plugin:runIde" />
         </list>
       </option>
       <option name="vmOptions" />
