+++ README.md
@@ -1,6 +1,6 @@
-Json2Kotlin
+Kotlinify
 =====
-[![Build Project](https://github.com/t28hub/json2kotlin/actions/workflows/build.yml/badge.svg)](https://github.com/t28hub/json2kotlin/actions/workflows/build.yml)
+[![Build Project](https://github.com/t28hub/kotlinify/actions/workflows/build.yml/badge.svg)](https://github.com/t28hub/kotlinify/actions/workflows/build.yml)
 
 ## License
