+++ gradle.properties
@@ -1,0 +1,23 @@
+#
+# Copyright (c) 2022 Tatsuya Maki
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+kotlin.code.style=official
+org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
+
+plugin.group=io.t28.kotlinify
+plugin.name=kotlinify
+plugin.version=1.0.0-SNAPSHOT
+
+gradle.version=7.3
