+++ kotlinify-idea-plugin/src/main/resources/META-INF/plugin.xml
@@ -1,0 +1,54 @@
+<!--
+  ~ Copyright (c) 2022 Tatsuya Maki
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~         http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+<idea-plugin>
+  <id>io.t28.kotlinify</id>
+  <name>Kotlinify</name>
+  <vendor email="t28oogle@gmail.com" url="https://github.com/t28hub">Tatsuya Maki</vendor>
+  <resource-bundle>messages.KotlinifyBundle</resource-bundle>
+
+  <description>
+    <![CDATA[
+        JSON to Kotlin data classes conversion plugin for IntelliJ IDEA.
+        ]]>
+  </description>
+
+  <!-- See https://plugins.jetbrains.com/docs/intellij/plugin-compatibility.html#modules-available-in-all-products -->
+  <depends>com.intellij.modules.platform</depends>
+  <depends>com.intellij.modules.lang</depends>
+  <depends>com.intellij.modules.java</depends>
+  <depends>org.jetbrains.kotlin</depends>
+
+  <extensions defaultExtensionNs="com.intellij">
+    <notificationGroup
+      id="io.t28.kotlinify.idea"
+      key="notification.group"
+      bundle="messages.KotlinifyBundle"
+      displayType="BALLOON"
+      toolWindowId="Project"
+    />
+  </extensions>
+
+  <actions>
+    <action
+      id="Kotlinify.New"
+      class="io.t28.kotlinify.idea.NewKotlinFileAction"
+      text="Kotlin Class/File from JSON"
+      description="Generate new Kotlin classes from JSON"
+    >
+      <add-to-group group-id="NewGroup" anchor="after" relative-to-action="Kotlin.NewFile" />
+    </action>
+  </actions>
+</idea-plugin>
