+++ kotlinify-idea-plugin/src/main/resources/messages/KotlinifyBundle.properties
@@ -1,0 +1,40 @@
+#
+# Copyright (c) 2022 Tatsuya Maki
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+plugin.name=Kotlinify
+
+# Dialog components
+action.new.file.dialog.title=New Kotlin Class/File from JSON
+action.new.file.dialog.name.title=Class name:
+action.new.file.dialog.format.title=Format:
+action.new.file.dialog.format.json=JSON
+action.new.file.dialog.format.jsonschema=JSON Schema
+action.new.file.dialog.text.title=Source text:
+
+## Buttons
+action.new.file.dialog.button.ok=Generate
+action.new.file.dialog.button.cancel=Cancel
+action.new.file.dialog.button.format=Format
+
+## Error messages
+action.new.file.dialog.error.name.empty=Class name is empty
+action.new.file.dialog.error.name.invalid=Class name '%s' contains invalid chars
+action.new.file.dialog.error.text.empty=Input text is empty
+action.new.file.dialog.error.text.unparseable=Input text is not parseable
+action.new.file.dialog.error.text.invalid.json=Input text is not JSON object or JSON array
+action.new.file.dialog.error.text.invalid.jsonschema=Input text is invalid JSON Schema
+
+# Notification
+notification.group=Kotlinify
