--- src/main/kotlin/io/t28/json2kotlin/idea/util/AnActionEvent.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2022 Tatsuya Maki
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.t28.json2kotlin.idea.util
-
-import com.intellij.ide.IdeView
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.actionSystem.LangDataKeys
-
-/**
- * Retrieve [IdeView] if exists
- */
-val AnActionEvent.ideView: IdeView?
-    get() {
-        return getData(LangDataKeys.IDE_VIEW)
-    }
